Update buffer to 1 GB
So the reason I'm dropping from 4gb -> 1 GB is because I really do not want any strong hardware constraints on operating the tool in different environments that may have limited memory allocated.

The allocation is also used for any file reading. Despite a bigger buffer being faster, its not typical enough to drastically change the program and is a more niche use case.