V2.0.0 (#57)
Restoring `v1.0.0` as `v2.0.0`.

### Why?
In `v1.1.0` I performed a large reduction in the logic to remove the dependence on loading items into memory. Prior tools operated like this, but `ptt` was actually designed to use in-map memory structures because it was designed for lighter operations than its predecessors. This created a program that worked better at some things, but largely had a lot less functionality.

The question is: Where do tools like this fit into the world today? Plenty of tools are now per user and more people than ever are making their own tools, which, I think, is great. Many might be searching for inspiration and structure rather than specific features, with this in mind, I think `v1.0.0` is a much better project state than where the current version sits, so reverting makes sense.

The tool will be a great inspiration, a functional tool, aligned with a large amount of documentation, and stable for years to come.